feat: works in browser
BREAKING CHANGE: Now uses EventEmitter3 for events in order to work in browsers, and name changes of exports. `SourcedEntity` is now `Entity` and a named export instead of a default export. `SourcedValue` was removed. `Value` is now a named export. `Entity` is now `EntityProxy`. Removed dependency on `util` to work in the browser.

Signed-off-by: Patrick Lee Scott <pat@patscott.io>